Add comprehensive unit tests for Pulse Width calculations (#1267)
* Refactor: extract and unit test pwApplyNitrous()

* Enhance calculatePWLimit unit tests

* Extract percentApprox from PW calculations, optimize and unit test.

* Enhance PW() unit tests

* Add nitrous adjustment inside PW()

* Calculate required fuel dynamically

* percentage() -  allow percent>UINT8_MAX

* calculateStaging - Inject all dependencies

* Extract & unit test applyPwLimit()

* Separate *calcuating* the staging PWs from *applying* them and unit test both functions

* Add fastDiv()

* Remove pre-calculated staging percentages - no loss in speed!

* Move calculateSecondaryPw() & applyPwToInjectorChannels() into pw_calcs.cpp

* Separate out & unit test calculateOpenTime

* Move calculateOpenTime() into fuel_calcs

* Move revolutionTime into statuses - will shorten parameter lists

* Bring PW calc orchestration togther into computePulseWidths()

* Use a realstic REQ_FUEL value - it must be a multiple of 50

* computePulseWidths(): inject all depedencies & add unit tests.
This makes test_PW.cpp redundant, so remove it.

* Deep unit test of PW() and rename to calcPrimaryPulseWidth

* Move maxIgnOutputs & maxInjOutputs into statuses struct

* Simplify: assign staging PWs purely based on having a secondary PW

* No need to check maxInjOutputs - the channel PW will be zero if maxInjOutputs is too low

* Add zero correction test

* Micro optimizations

* Bug Fix: nitrous percentage calculation can overflow int16_t (and uint16_t) if the RPM delta is high enough

* Increase accuracy of nitrous PW calculation (was previously rounding to 100ms).

* Remove unused macro

* MISRA fixes

* Clean code: gather fuel schedule configuration into one function (setFuelSchedules); use an array for injection start angles

* Remove globals.h from fuel_calcs

* Remoev ARM specific overloads

+++ speeduino/secondaryTables.cpp
@@ -146,7 +146,7 @@ void calculateSecondarySpark(const config2 &page2, const config10 &page10, const
       current.secondSparkTableActive = true;
       uint8_t spark2Percent = (uint8_t)clamp(lookupSpark2(page10, sparkLookupTable, current), (int16_t)0, (int16_t)UINT8_MAX);
       //Spark 2 table is treated as a % value. Table 1 and 2 are multiplied together and divided by 100
-      int16_t combinedAdvance = div100((int16_t)spark2Percent * (int16_t)current.advance1);
+      int16_t combinedAdvance = div100((int16_t)(spark2Percent * current.advance1));
       //make sure we don't overflow and accidentally set negative timing: current.advance can only hold a signed 8 bit value
       current.advance = constrainAdvance(combinedAdvance);
